requirements.txt
opencv-python
ultralytics
geocoder
python-dotenv
google-generativeai
Pillow
torch
